# Speedup fasta export