--- a/grails-app/services/org/bbop/apollo/SequenceService.groovy
+++ b/grails-app/services/org/bbop/apollo/SequenceService.groovy
@@ -207,44 +207,14 @@ class SequenceService {
         int startChunkNumber = fmin / sequence.seqChunkSize;
         int endChunkNumber = (fmax - 1 ) / sequence.seqChunkSize;
 
-        def c=SequenceChunk.createCriteria()
-        def chunks=c.list() {
-            eq('sequence',sequence)
-            'in'('chunkNumber',startChunkNumber..endChunkNumber)
-            order("chunkNumber", "asc")
+        
+        for(int i = startChunkNumber ; i<= endChunkNumber ; i++){
+            sequenceString.append(loadResidueForSequence(sequence,i))
         }
-        log.debug("${chunks.size()} ${endChunkNumber-startChunkNumber+1}")
-        if(chunks.size()==endChunkNumber-startChunkNumber+1) {
-            chunks.each {
-                sequenceString.append(it.residue)
-            }
-        }
-        else {
-            for(int i = startChunkNumber ; i<= endChunkNumber ; i++){
-                SequenceChunk sequenceChunk = getSequenceChunkForChunk(sequence,i)
-                sequenceString.append(sequenceChunk.residue)
-            }
-        }
-
 
         int startPosition = fmin - (startChunkNumber * sequence.seqChunkSize);
 
         return sequenceString.substring(startPosition,startPosition + (fmax-fmin))
-    }
-
-    SequenceChunk getSequenceChunkForChunk(Sequence sequence, int i) {
-        SequenceChunk sequenceChunk = SequenceChunk.findBySequenceAndChunkNumber(sequence,i)
-        if(!sequenceChunk){
-            String residue = loadResidueForSequence(sequence,i)
-            log.debug "RESIDUE load: ${residue?.size()}"
-            sequenceChunk = new SequenceChunk(
-                    sequence: sequence
-                    ,chunkNumber: i
-                    ,residue: residue
-            ).save(flush:true)
-        }
-        log.debug "RESIDUE loaded from DB: ${sequenceChunk.residue?.size()}"
-        return sequenceChunk
     }
 
     private static String generatorSampleDNA(int size){