@nathandunn 2016-02-05
Interesting, but not suprising that the chunk code wasn't doing anything helpful.  This was adopted from what Ed had done originally in-memory, but is obviously not a pinch-point here.