@nathandunn 2016-02-05
for 50 genes: 15 seconds -> 1.7 seconds . . .  I tested both FASTA (all types) and GFF + FASTA . . . . looked good to me